+++ SendGrid/SendGrid/Client.cs
@@ -15,6 +15,9 @@ namespace SendGrid
         private string _apiKey;
         public APIKeys ApiKeys;
         public UnsubscribeGroups UnsubscribeGroups;
+        public Suppressions Suppressions;
+        public GlobalSuppressions GlobalSuppressions;
+        public GlobalStats GlobalStats;
         public string Version;
         private Uri _baseUri;
         private const string MediaType = "application/json";
@@ -35,6 +38,9 @@ namespace SendGrid
             Version = Assembly.GetExecutingAssembly().GetName().Version.ToString();
             ApiKeys = new APIKeys(this);
             UnsubscribeGroups = new UnsubscribeGroups(this);
+            Suppressions = new Suppressions(this);
+            GlobalSuppressions = new GlobalSuppressions(this);
+            GlobalStats = new GlobalStats(this);
         }
 
         /// <summary>