+++ SendGrid/UnitTest/UnitTest.cs
@@ -156,4 +156,127 @@ namespace UnitTest
         }
 
     }
+
+    [TestFixture]
+    public class Suppressions
+    {
+        static string _baseUri = "https://api.sendgrid.com/";
+        static string _apiKey = Environment.GetEnvironmentVariable("SENDGRID_APIKEY");
+        public Client client = new Client(_apiKey, _baseUri);
+
+        [Test]
+        public void SuppressionsIntegrationTest()
+        {
+            int unsubscribeGroupId = 69;
+
+            TestGet(unsubscribeGroupId);
+            string[] emails = { "example@example.com", "example2@example.com" };
+            TestPost(unsubscribeGroupId, emails);
+            TestDelete(unsubscribeGroupId, "example@example.com");
+            TestDelete(unsubscribeGroupId, "example2@example.com");
+        }
+
+        private void TestGet(int unsubscribeGroupId)
+        {
+            HttpResponseMessage response = client.Suppressions.Get(unsubscribeGroupId).Result;
+            Assert.AreEqual(HttpStatusCode.OK, response.StatusCode);
+            string rawString = response.Content.ReadAsStringAsync().Result;
+            dynamic jsonObject = JsonConvert.DeserializeObject(rawString);
+            Assert.IsNotNull(jsonObject);
+        }
+
+        private void TestPost(int unsubscribeGroupId, string[] emails)
+        {
+            HttpResponseMessage response = client.Suppressions.Post(unsubscribeGroupId, emails).Result;
+            Assert.AreEqual(HttpStatusCode.Created, response.StatusCode);
+            string rawString = response.Content.ReadAsStringAsync().Result;
+            dynamic jsonObject = JObject.Parse(rawString);
+            string recipient_emails = jsonObject.recipient_emails.ToString();
+            Assert.IsNotNull(recipient_emails);
+        }
+
+        private void TestDelete(int unsubscribeGroupId, string email)
+        {
+            HttpResponseMessage response = client.Suppressions.Delete(unsubscribeGroupId, email).Result;
+            Assert.AreEqual(HttpStatusCode.NoContent, response.StatusCode);
+        }
+
+    }
+
+    [TestFixture]
+    public class GlobalSuppressions
+    {
+        static string _baseUri = "https://api.sendgrid.com/";
+        static string _apiKey = Environment.GetEnvironmentVariable("SENDGRID_APIKEY");
+        public Client client = new Client(_apiKey, _baseUri);
+
+        [Test]
+        public void GlobalSuppressionsIntegrationTest()
+        {
+            string email = "example3@example.com";
+
+            TestGet(email);
+            string[] emails = { "example1@example.com", "example2@example.com" };
+            TestPost(emails);
+            TestDelete("example1@example.com");
+            TestDelete("example2@example.com");
+        }
+
+        private void TestGet(string email)
+        {
+            HttpResponseMessage response = client.GlobalSuppressions.Get(email).Result;
+            Assert.AreEqual(HttpStatusCode.OK, response.StatusCode);
+            string rawString = response.Content.ReadAsStringAsync().Result;
+            dynamic jsonObject = JsonConvert.DeserializeObject(rawString);
+            Assert.IsNotNull(jsonObject);
+        }
+
+        private void TestPost(string[] emails)
+        {
+            HttpResponseMessage response = client.GlobalSuppressions.Post(emails).Result;
+            Assert.AreEqual(HttpStatusCode.Created, response.StatusCode);
+            string rawString = response.Content.ReadAsStringAsync().Result;
+            dynamic jsonObject = JObject.Parse(rawString);
+            string recipient_emails = jsonObject.recipient_emails.ToString();
+            Assert.IsNotNull(recipient_emails);
+        }
+
+        private void TestDelete(string email)
+        {
+            HttpResponseMessage response = client.GlobalSuppressions.Delete(email).Result;
+            Assert.AreEqual(HttpStatusCode.NoContent, response.StatusCode);
+        }
+    }
+
+    [TestFixture]
+    public class GlobalStats
+    {
+        static string _baseUri = "https://api.sendgrid.com/";
+        static string _apiKey = Environment.GetEnvironmentVariable("SENDGRID_APIKEY");
+        public Client client = new Client(_apiKey, _baseUri);
+
+        [Test]
+        public void GlobalStatsIntegrationTest()
+        {
+            string startDate = "2015-11-01";
+            string endDate = "2015-12-01";
+            string aggregatedBy = "day";
+            TestGet(startDate);
+            TestGet(startDate, endDate);
+            TestGet(startDate, endDate, aggregatedBy);
+            aggregatedBy = "week";
+            TestGet(startDate, endDate, aggregatedBy);
+            aggregatedBy = "month";
+            TestGet(startDate, endDate, aggregatedBy);
+        }
+
+        private void TestGet(string startDate, string endDate=null, string aggregatedBy=null)
+        {
+            HttpResponseMessage response = client.GlobalStats.Get(startDate, endDate, aggregatedBy).Result;
+            Assert.AreEqual(HttpStatusCode.OK, response.StatusCode);
+            string rawString = response.Content.ReadAsStringAsync().Result;
+            dynamic jsonObject = JsonConvert.DeserializeObject(rawString);
+            Assert.IsNotNull(jsonObject);
+        }
+    }
 }