+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 ﻿# Change Log
 All notable changes to this project will be documented in this file.
 
+## [6.3.4] - 2015-12-15
+###Added
+- Implemented the global stats /asm/stats endpoint [GET]
+
+## [6.3.3] - 2015-12-14
+###Added
+- Implemented the global suppressions /asm/suppressions/global endpoint [GET, POST, DELETE]
+
+## [6.3.2] - 2015-12-11
+###Added
+- Implemented the suppressions /asm/groups/:group_id/suppressions endpoint [GET, POST, DELETE]
+
 ## [6.3.1] - 2015-12-10
 ###Added
 - Implemented the unsubscribe groups /asm/groups endpoint [GET, POST, DELETE]